lector-resumenes-bancarios/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
et_xmlfile==2.0.0
fastapi==0.135.2
h11==0.16.0
idna==3.11
openpyxl==3.1.5
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079150/work
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.6.0
pytesseract==0.3.13
python-multipart==0.0.22
setuptools==80.10.2
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
wheel==0.46.3

